switch to 4-wide f32 SIMD and add example render (#2)
* Remove fake dependencies

I had in mind to make this a gpu black hole raytracer, and forgot to remove them.

* Use 4-wide f32 vectors, drop trig randomUnit

Switched Vec3 from @Vector(3, f64) to @Vector(4, f32) to actually use SIMD. Replaced randomUnit with rejection sampling, and seed the per thread PRNG deterministically so threads no longer share a single random stream. Dropped some dead code.

Saved about ~3 seconds on the render settings I used as benchark. Bit of a waste of time.

* set optimized float mode

Added @setFloatMode(.optimized) to the places I though it would make a diffrence.

Saved ~3.5 seconds from render. So, that's ~6.5 shaved from render time for the benchmark. Progress!

* add example render

+++ src/color.zig
@@ -7,7 +7,7 @@ pub const Color = Vec3;
 
 const Interval = @import("interval.zig").interval;
 
-inline fn linearToGamma(linearComponent: f64) f64 {
+inline fn linearToGamma(linearComponent: f32) f32 {
     if (linearComponent > 0.0) {
         return std.math.sqrt(linearComponent);
     }
@@ -15,9 +15,9 @@ inline fn linearToGamma(linearComponent: f64) f64 {
 }
 
 pub fn colorToBytes(pixel_color: Color) [3]u8 {
-    var r: f64 = pixel_color[0];
-    var g: f64 = pixel_color[1];
-    var b: f64 = pixel_color[2];
+    var r: f32 = pixel_color[0];
+    var g: f32 = pixel_color[1];
+    var b: f32 = pixel_color[2];
 
     r = linearToGamma(r);
     g = linearToGamma(g);
