+++ src/interval.zig
@@ -3,10 +3,10 @@ const std = @import("std");
 const rtweekend = @import("rtweekend.zig");
 
 pub const interval = struct {
-    min: f64,
-    max: f64,
+    min: f32,
+    max: f32,
 
-    pub fn init(min: f64, max: f64) interval {
+    pub fn init(min: f32, max: f32) interval {
         return interval{ .min = min, .max = max };
     }
 
@@ -14,15 +14,15 @@ pub const interval = struct {
         return interval{ .min = rtweekend.infinity, .max = -rtweekend.infinity };
     }
 
-    pub fn size(self: interval) f64 {
+    pub fn size(self: interval) f32 {
         return self.max - self.min;
     }
 
-    pub fn contains(self: interval, x: f64) bool {
+    pub fn contains(self: interval, x: f32) bool {
         return self.min <= x and x <= self.max;
     }
 
-    pub fn surrounds(self: interval, x: f64) bool {
+    pub fn surrounds(self: interval, x: f32) bool {
         return self.min < x and x < self.max;
     }
 
@@ -30,7 +30,7 @@ pub const interval = struct {
         return interval{ .min = -rtweekend.infinity, .max = rtweekend.infinity };
     }
 
-    pub fn clamp(self: interval, x: f64) f64 {
+    pub fn clamp(self: interval, x: f32) f32 {
         if (x < self.min) return self.min;
         if (x > self.max) return self.max;
